# openapi.yaml
openapi: 3.1.0
info:
  title: Veilence-MX API
  description: >
    Enterprise SaaS supply chain compromise monitor for PyPI and npm packages.
    Provides endpoints for package monitoring, release analysis, alert management,
    organization/RBAC, notifications, and system administration.
  version: 1.0.0
  contact:
    name: Veilence-MX Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.veilence.example.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/packages:
    get:
      tags: [Packages]
      summary: List packages
      description: Returns a paginated list of monitored packages scoped to the current organization.
      operationId: listPackages
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SortBy"
        - $ref: "#/components/parameters/SortDir"
        - name: registry
          in: query
          schema:
            type: string
            enum: [pypi, npm]
        - name: is_custom
          in: query
          schema:
            type: string
            enum: ["true", "false"]
        - name: search
          in: query
          schema:
            type: string
          description: Search by package name (ILIKE)
      responses:
        "200":
          description: Paginated package list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_PackageList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Packages]
      summary: Add package to monitoring
      description: Adds a custom package to monitoring. Requires `packages:write` permission.
      operationId: createPackage
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageRequest"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Package"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/packages/{id}:
    get:
      tags: [Packages]
      summary: Get package
      description: Returns a single package by ID.
      operationId: getPackage
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Package details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Package"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Packages]
      summary: Delete package
      description: Removes a package from monitoring. Requires `packages:delete` permission.
      operationId: deletePackage
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Package deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login or /api/auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key created via /api/auth/api-keys
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    XOrgId:
      name: X-Org-ID
      in: header
      required: true
      description: Organization ID (can also be passed as `org_id` query parameter)
      schema:
        type: integer
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortBy:
      name: sort_by
      in: query
      schema:
        type: string
      description: Column to sort by (varies per endpoint)
    SortDir:
      name: sort_dir
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        data:
          nullable: true
        error:
          type: string

    # ── Domain Schemas ──────────────────────────────────────────────
...
    Package:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        name:
          type: string
        registry:
          type: string
          enum: [pypi, npm]
        latestVersion:
          type: string
        description:
          type: string
        isCustom:
          type: boolean
        rank:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
          format: int64

    # ── Request Schemas ─────────────────────────────────────────────
...
    CreatePackageRequest:
      type: object
      required: [name, registry]
      properties:
        name:
          type: string
        registry:
          type: string
          enum: [pypi, npm]
...
    ApiResponse_Package:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Package"
        error:
          type: string
          nullable: true
    ApiResponse_PackageList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Package"
        error:
          type: string
          nullable: true
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...